requirements.txt
pandas
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
ensure==1.0.4
joblib
types-pyYAML
Flask
Flask-Cors
streamlit
psycopg2
-e.